Execute a single stratum or lamina ad hoc by its name

In interactive sessions, adhoc() will execute the stratum or lamina that matches the name provided by the user. If multiple matches are found, the user will be prompted to choose which one to execute. If no matches are found, an error will be thrown. project_path will default to the current working directory, unless a path is provided by the user.

Usage

adhoc(name, prompt = TRUE, silent = FALSE, project_path = NULL)

Arguments

name

Name of stratum or lamina.

prompt

Prompt user for choice if multiple matches found? Default is TRUE.

silent

Suppress log messages? If FALSE (the default), log messages will be printed to the console. If TRUE, log messages will be suppressed.

project_path

A path to strata project folder.

Value

invisible data frame of execution plan for matched name.

See also

Examples

if (FALSE) { # \dontrun{
tmp <- fs::dir_create(fs::file_temp())
result <- strata::build_quick_strata_project(tmp, 3, 2)
adhoc("stratum_1")
fs::dir_delete(tmp)
} # }