Quickly build strata project with minimal input and standard names

Quickly build strata project with minimal input and standard names

Usage

build_quick_strata_project(project_path, num_strata = 1, num_laminae_per = 1)

Arguments

project_path

Path to the project folder

num_strata

Number of strata to create

num_laminae_per

Number of laminae to create per stratum

Value

invisible dataframe of the survey of the strata project

Examples

tmp <- fs::dir_create(fs::file_temp())
result <- build_quick_strata_project(tmp, 2, 2)
dplyr::glimpse(result)
#> Rows: 4
#> Columns: 7
#> $ execution_order <int> 1, 2, 3, 4
#> $ stratum_name    <chr> "stratum_1", "stratum_1", "stratum_2", "stratum_2"
#> $ lamina_name     <chr> "s1_lamina_1", "s1_lamina_2", "s2_lamina_1", "s2_lamin…
#> $ script_name     <chr> "my_code", "my_code", "my_code", "my_code"
#> $ script_path     <fs::path> "/tmp/Rtmpvy72x9/file16dc62407d20/strata/stratum_1/s1_…
#> $ skip_if_fail    <lgl> FALSE, FALSE, FALSE, FALSE
#> $ created         <date> 2024-11-19, 2024-11-19, 2024-11-19, 2024-11-19