Send a standardized log message to stdout or stderr

log_message() does not stop the execution of the script, regardless of the level of the message, and whether or not it prints to STDOUT or STDERR.

Usage

log_message(message, level = "INFO", out_or_err = "OUT")

Arguments

message

A string containing a message to log.

level

The level of the message (e.g. INFO, WARNING, ERROR), defaults to "INFO" but will accept any string.

out_or_err

Send log output to stdout or stderr, choices are "OUT" or "ERR" and the defaults is "OUT".

Value

A message printed to stdout or stderr and an invisible character string copy of the entire log message.

See also

Examples

log_message("This is an info message", "INFO", "OUT")
#> [2025-03-05 13:50:30.1026] INFO: This is an info message 
log_message("This is an error message", "ERROR", "ERR")
#> [2025-03-05 13:50:30.1036] ERROR: This is an error message
log_message("This is a warning message", "WARNING", "OUT")
#> [2025-03-05 13:50:30.1046] WARNING: This is a warning message