Survey the layout and execution order of your project

survey_strata() will examine the .tomls in project_path provided and return a dataframe with the following information about the project:

stratum_name: the name of the stratum

lamina_name: the name of the lamina

execution_order: the order in which the stratum-lamina-code combination will be executed

script_name: the name of the script to be executed

script_path: the path to the script

This is based on the contents of the .toml files, everything else is "invisible" inside the strata project.

Usage

survey_strata(project_path)

Arguments

project_path

A path to strata project folder.

Value

dataframe housing the layout of your project based on the .tomls.

Examples

tmp <- fs::dir_create(fs::file_temp())
build_quick_strata_project(tmp, 2, 2)
survey_strata(tmp)
#> # A tibble: 4 × 7
#>   execution_order stratum_name lamina_name script_name script_path  skip_if_fail
#>             <int> <chr>        <chr>       <chr>       <fs::path>   <lgl>       
#> 1               1 stratum_1    s1_lamina_1 my_code     …1/my_code.R FALSE       
#> 2               2 stratum_1    s1_lamina_2 my_code     …2/my_code.R FALSE       
#> 3               3 stratum_2    s2_lamina_1 my_code     …1/my_code.R FALSE       
#> 4               4 stratum_2    s2_lamina_2 my_code     …2/my_code.R FALSE       
#> # ℹ 1 more variable: created <date>
fs::dir_delete(tmp)